Fixes boundingNodes sorting logic
boundingNodes was using `indexPath` which was a string.

Not only was this a problem for deep nodes because it turned the string
into one giant string which sorted as a whole not as each nesting level
but it would also sort different levels wrong - e.g. 11 comes before 5.

This turns it into a numeric sort that considers each nesting level.